### 2020-07-18 #86 @whedon
[PRE REVIEW]: Course Materials for Data Science in Practice
**Submitting author:** @TomDonoghue (<a href="http://orcid.org/0000-0001-5911-0472">Thomas Donoghue</a>)
**Repository:** <a href="https://github.com/DataScienceInPractice/Site" target ="_blank">https://github.com/DataScienceInPractice/Site</a>
**Version:** 1.0

@TomDonoghue if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
