### 2020-07-18 #86 @labarba
👋  @TomDonoghue  — I'm ready to look for an editor to handle review of your submission. The repository you provided here seems to hold just the paper and some additional info, however. We expect the repository of the complete learning materials, with the JOSE paper either in a folder in there, or if you prefer visible in a branch. Could you reorganize your submission in this manner and report the GitHub URL here?

At the end of review, and when the paper is accepted, we will ask you to make an archive of the full repository on Zenodo or a similar service, and the archive DOI will be linked in the paper and added in the metadata of the paper's Crossreff deposit.
