### 2020-07-22 #86 @TomDonoghue
Hey @labarba - thank you handling the submission!

In terms of an editor, I wasn't sure about recommending a specific editor at submission (in terms of who is available in these times), but in terms of content, when I checked the listing, potentially @\juanklopper is one of the most related, by topic.

In term of the repository, the project collects and makes public some materials for a university course, and the main "deliverable" is the website with all of the materials (https://datascienceinpractice.github.io/). I have now updated the repository so that the repository itself also includes the materials that are hosted. I have also updated the README to better describe the repository and it's contents. Hopefully this matches what is needed for the JOSE submission, and if any further updates are required, please let me know!

Please let us know if there is anything else we can do / assist with for the submission!
