+++ dsip/projects/projects.md
@@ -1,0 +1,177 @@
+# Project Description
+
+## Project Overview
+
+The Final Project will give you the chance to explore a topic of your choice and to expand your analytical skills. By working with real data of your choosing you can examine questions of particular interest to you.
+
+The broad objectives for the project are to:
+
+* Identify the problems and goals of a real situation and dataset.
+* Choose an appropriate approach for formalizing and testing the problems and goals, and be able to articulate the reasoning for that selection.
+* Implement your analysis choices on the dataset.
+* Interpret the results of the analyses.
+* Contextualize those results within a greater scientific and social context, acknowledging and addressing any potential issues related to privacy and ethics.
+* Work effectively to manage a project as part of a team.
+
+To accomplish this you will work in teams of 4 to 5 students to conceive of and carry out an analysis project. ***Everyone who chooses this option must be part of a group.*** You will find in your future careers the need to work on projects in groups frequently (even if you really, really, really, really don’t want to).
+
+The basic project steps:
+
+* Find a real world dataset and problem that you believe can be solved with one or more of the techniques we have learned in class.
+* After selecting a dataset and identifying the goal, write out a proposed analysis plan using template provided and submit it through GitHub for review.
+* Apply the techniques outlined and come up with a result for the dataset that you proposed.
+* Assemble a Jupyter notebook that communicates your hypothesis, methods, and results. Submit this as your final project.
+* Submit feedback about group and individual group members. This is done individually.
+
+## Getting Started
+
+We strongly encourage you to discuss potential project ideas on Piazza, with your TAs and IAs, and/or with Prof. Ellis! This will give us a chance to make sure you’re on the right track even before you submit your draft.
+
+## The Project Proposal
+
+**RESEARCH QUESTION**: What is your research question? Include the specific question you're setting out to answer. (1-2 sentences)
+
+**BACKGROUND & PRIOR WORK**: It will present the background and context of your topic and question in a few paragraphs. Describe what information you all currently know about the topic, include references to other projects who have asked similar questions or approached similar problems. Explain what others have learned in their projects.
+
+* Why is this question of interest to your group?
+* What background information led you to your hypothesis.
+* Why is this important?
+* What has already been done on this topic? What is already known?
+
+Find some relevant prior work, and reference those sources. Even if you think you have a totally novel question, find the most similar prior work that you can and discuss how it relates to your project.
+
+References can be research publications, but they need not be. Blogs, GitHub repositories, company websites, etc., are all viable references if they are relevant to your project. It must be clear which information comes from which references. (2-3 paragraphs, including at least 2 references)
+
+**HYPOTHESIS**: What is your main hypothesis and predictions? Briefly explain why. (2-3 sentences)
+
+**DATA**: Here, you are to *think* about  the *ideal* dataset (or datasets) you you would need to answer this question. Describe and explain the ideal datasets you would want to answer your question.
+
+* What variables would you have?
+* How would they be stored?
+* How many observations would you have?
+* What/who would the observations be? etc.)
+* etc.
+
+Note: For the project proposal, you do not have to find the actual dataset(s) needed for your project. For your final project, you will.
+
+**ETHICS & PRIVACY**: Acknowledge and address any ethics & privacy related issues of your question(s), proposed dataset(s), and/or analyses. Use the information provided in lecture to guide your group discussion and thinking. If you need further guidance, check out [Deon's Ethics Checklist](http://deon.drivendata.org/#data-science-ethics-checklist). In particular:
+
[…]
