### 2020-08-14 #86 @TomDonoghue
Hey @labarba - I just wanted to check in - is there anything more we can do to prepare the repository submission, and/or to help assign a handling editor?

I'm not sure if we are supposed to directly link in editors, but I think potentially @\juanklopper is related to the topic, and/or @\magsol does related work (using Python for data analysis, etc).
