### 2020-09-06 #86 @kyleniemeyer
Hi @TomDonoghue, sorry for the extended delay here!

@juanklopper, are you able to take on this paper as editor?
