### 2020-11-21 #86 @TomDonoghue
Pinging on this - it seems perhaps @magsol  is not available to handle the submission?

If not, @kyleniemeyer and/or @labarba - could you let you know if JOSE is operational and if you expect this submission to be processed / reviewed, and if so, when?

I understand things are slower during a pandemic - but we submitted this paper around July 1st - almost 5 months ago. There was not then (nor is there now) any indication at submission JOSE was not actively processing papers. If we had JOSE was not operational, we would not have spent time preparing this submission. Please let us know what is going on here, and if things are to remain paused - I would suggest at least mentioning this on the JOSE journal & submission page, so that potential submitters know the status.
