### 2020-11-21 #86 @labarba
Hi @TomDonoghue — It is I who became the bottleneck for JOSE, as the pandemic made all my service obligations too much to do them all. I was JupyterCon General Chair, and the pivot to an online event was a huge lift (it was held in October). I'm just beginning to dig from under the rubble of all that I'm behind on, and will start processing the several stalled JOSE submissions. Thank you for your patience.
