### 2020-11-21 #86 @magsol
Hi @TomDonoghue and @kyleniemeyer, so sorry I missed the earlier pings--I've been on paternity leave the last several weeks so quite a lot has fallen through the cracks. I'm still trying to push through a separate stalled JOSE submission, but if you can't find another editor, I can see about this one.
