### 2020-11-25 #86 @TomDonoghue
@labarba - thanks for getting back to us and the update. I understand there is a lot going on! Is there anything we can do to help move this paper along?

In terms of assigning a managing editor, my original editor suggestions were @juanklopper and/or @magsol (thanks for answering - but I don't want to pull you in from paternity leave, if avoidable!). The submission is fairly broad (in terms of covering data science / scientific Python), so I imagine other editors would overlap a lot with it, in so far as they work within scientific python - so other possible editors might be @\moorepants or @\ianhawke, based on my understanding of the editor list.
