### 2021-03-10 #86 @TomDonoghue
Trying again.... @labarba and/or @kyleniemeyer - do you think it would be possible to get this going? We would still like to get this reviewed and potentially published in JOSE if possible. We still need a managing editor.
