### 2021-06-28 #86 @TomDonoghue
@magsol - sounds good, thanks for jumping on board! It's been a little while since I worked through the details of this paper, but everything should be in pretty good shape I think - the paper generates, and refers to a collection of online learning materials available one the website.

The topic is pretty general data science (in Python) - do you need me to look through the JOSE reviewer list and recommend some potential reviewers?
