### 2021-06-29 #86 @TomDonoghue
Hey @magsol: in terms of reviewers, I checked through the list of reviewers, checking through based on the intersection of Data Science + Python (+teaching and/or other relevant topics), and I find the following potential reviewers (non exhaustive list):
- Dr. Kristian Rother (krother)
- Sarah Masud (sara-02)
- Albert Y. Kim (rudeboybert)
- Giulio Valentino Dalla Riva (gvdr)
- Justin Bois (justinbois)
- Tania Allard (trallard)
- Brandon LeBeau (lebebr01)

Also, assuming it's okay to do so, in the next day or two, I'll do a sweep through the paper, and check if there are any minor updates to do (since it's been a while since I looked at this paper), and add in a reference to some other work we've done on teaching data science.
