### 2021-07-06 #121 @whedon
[REVIEW]: Course Materials for Data Science in Practice
**Submitting author:** @TomDonoghue (<a href="http://orcid.org/0000-0001-5911-0472">Thomas Donoghue</a>)
**Repository:** <a href="https://github.com/DataScienceInPractice/Site" target ="_blank">https://github.com/DataScienceInPractice/Site</a>
**Version:** v1.1

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
