### 2021-07-14 DataScienceInPractice/Site#4 @krother
Missing release / version number for review
On https://github.com/openjournals/jose-reviews/issues/121 a release is mentioned as part of the review checklist. I suggest creating one on https://github.com/DataScienceInPractice/Site/tags and labeling it `v1.0` or similar.
