### 2021-07-14 #121 @krother
@KiCh84 I created an issue for the missing release, so that the final paper refers to a defined state of the repository.
