### 2021-07-14 DataScienceInPractice/Site#5 @krother
Phantom Python Introduction
When I click on **pandas** https://datascienceinpractice.github.io/assignments/D2_Pandas.html  there is no "Introduction to Python" in **Assignments** in the table of contents on the left.

When I click on **Data Analysis** https://datascienceinpractice.github.io/assignments/D5_DataAnalysis.html  it suddenly appears.

This issue was raised as part of the JOSE review https://github.com/openjournals/jose-reviews/issues/121
