### 2021-07-17 DataScienceInPractice/Site#6 @krother
Mismatching licenses in repository and JOSE paper
Hi @TomDonoghue ,

during the JOSE review https://github.com/openjournals/jose-reviews/issues/121 I discovered that your paper references the CC-BY 4.0 license, while the repository references CC-NC-BY 4.0. Both are approved OSI licenses, but I think they should be consistent.
