### 2021-07-27 #121 @TomDonoghue
Hey @krother - thanks a ton for the review! The things you brought up make total sense, I will fix them up soon.

@KiCh84 - have you had a chance to look at the project and do you have any issues / suggestions? If you have a timeline for adding any comments, I might try and group any updates together!
