### 2021-08-20 #121 @KiCh84
Dear Bradley,

I am reading through the course material, and I just quickly wanted to
check in and ask you about who the target audience is, for this wonderful
course ?
and I apologise heavily for the delay in the responses .

On Fri, Aug 6, 2021 at 7:43 PM Bradley Voytek ***@***.***>
wrote:

> @KiCh84 <https://github.com/KiCh84> no, rush since it sounds like a lot's
> happening (sorry!) but I wanted to see if you had any additional comments
> or recommendations for us!
>
