### 2021-08-20 #121 @voytek
No worries! The target is undergraduate students with some Python experience, and some statistics.

We have a complementary paper for that class that outlines the general target—and our motivations—[here](https://www.tandfonline.com/doi/full/10.1080/10691898.2020.1860725).

Thanks for the kind words, and I hope this clarifies!
