### 2021-10-01 #121 @labarba
@openjournals/dev — hi folks! I'm a little confused about the stats on this repo. What does it mean when the stats list 24k lines under "comment" for Jupyter Notebook? The repo has a lot of material that is not picked up by the stats tool?
