### 2021-10-01 #121 @arfon
> @openjournals/dev — hi folks! I'm a little confused about the stats on this repo. What does it mean when the stats list 24k lines under "comment" for Jupyter Notebook? The repo has a lot of material that is not picked up by the stats tool?

It looks like `cloc` is detecting 29 notebooks which seems about right. As for the actual LOC within, it's entirely possible that this tool doesn't know how to calculate LOC properly (i.e., is getting confused about what is code in a notebook).

Is there other material that you think it should be detecting?
