### 2021-10-01 #121 @labarba
> Is there other material that you think it should be detecting?

I don't know! It just threw me off when I looked at the `loc` and saw a small-ish number, then pulled the repo and saw all those Jupyter notebooks with code and all.
