### 2021-10-01 #121 @xuanxu
I suspect the tool is counting as 'Comments' every [cell](https://raw.githubusercontent.com/DataScienceInPractice/Site/main/dsip/tutorials/10-Distributions.ipynb) in the notebooks with `cell_type: "markdown"` (as opposed to `cell_type: code`)
