### 2021-10-14 #121 @krother
@magsol I created a couple of issues in the repository yet. The two related to the checkboxes are: the lack of a version number (not a big deal on my end). The other is that there are two different licenses in the repo and the paper, I believe this needs to be fixed.
