### 2021-11-02 #121 @voytek
Yeah, we can implement those changes pretty quickly. But we were waiting for all reviewers' comments. Should we just proceed?

cc @TomDonoghue
