### 2021-11-21 #121 @magsol
Hi folks! Where are we with this review? @krother @KiCh84 there are still unchecked items on both your lists, are those outstanding or have they been addressed yet? @TomDonoghue @voytek how does the feedback so far look?
