### 2022-01-16 #121 @magsol
Hi all! Touching base again.

@KiCh84 has your work situation improved at all? Any sort of timeline to completing the review? Or do we need to look for other reviewers?

@krother It looks like you're -almost- finished with the review; just a couple more items to check off. Are these items you can finish, or do you need feedback from the authors?

@TomDonoghue @voytek Any feedback for the reviewers?
