### 2022-02-15 #121 @krother
@magsol on my side, the two issues https://github.com/DataScienceInPractice/Site/issues/5 and https://github.com/DataScienceInPractice/Site/issues/6 are still open.
