### 2022-02-16 #121 @magsol
Thanks for the update, @krother .

@TomDonoghue @voytek Would you have a chance soon to look into those two remaining issues? If we can knock those out, then by that point hopefully we'll have heard back from @KiCh84 .
