### 2022-02-18 DataScienceInPractice/Site#5 @TomDonoghue
hey @krother - thanks for the note on this - I was able to replicate that on the previous deployment, there was some kind of quirk with the table of contents for the assignments, and in particular that `Python Introduction` would often disappear.

I have updated Jupyter-book, and redeployed, and no longer see this issue. It seems to be fixed with the re-deployment, so I'm going to close this issue - but please re-open if it doesn't seem to be fixed for you!
