### 2022-02-18 DataScienceInPractice/Site#6 @TomDonoghue
Hey @krother - thanks for the note, that was our bad on the discrepancy! I have updated and consolidated to use and list the CC-BY 4.0 license across the board.
