### 2022-02-20 47eec0c
add logo & tweak config
