+++ dsip/projects/project_guidelines.md
@@ -1,0 +1,337 @@
+
+# COGS 108 - Final Project
+
+**Table of Contents**
+
+- [Project Overview](#Project-Overview)
+  * [Taboo Topics](#Taboo-Topics)
+- [Project Components](#Project-Components)
+  * [Group Project Survey](#Group-Project-Survey)
+  * [Previous Project Review](#Previous-Project-Review)
+  * [Project Proposal](#Project-Proposal)
+  * [Checkpoints](#Checkpoints)
+  * [Final Project](#Final-Project)
+  * [Team Evaluation](#Team-Evaluation)
+- [Advice](#Advice)
+- [Previous COGS 108 Final Projects](#Previous-COGS-108-Final-Projects) 
+- [How to Find Datasets](#How-to-Find-Datasets)
+
+<!-- toc -->
+
+## Project Overview 
+
+The COGS 108 Final Project will give you the chance to explore a topic of your choice and to expand your analytical skills. By working with real data of your choosing you can examine questions of particular interest to you. 
+
+The broad objectives for the project are to: 
+
+* Identify the problems and goals of a real situation and dataset. 
+* Choose an appropriate approach for formalizing and testing the problems and goals, and be able to articulate the reasoning for that selection. 
+* Implement your analysis choices on the dataset. 
+* Interpret the results of the analyses. 
+* Contextualize those results within a greater scientific and social context, acknowledging and addressing any potential issues related to privacy and ethics. 
+* Work effectively to manage a project as part of a team. 
+
+To accomplish this you will work in teams of 4 to 5 students to conceive of and carry out an analysis project. You will find in your future careers the need to work on projects in groupcs frequently (even if you really, really, really, really don’t want to). 
+
+The basic project steps (broken down in more detail below): 
+
+* Find a real world dataset and problem that you believe can be solved with one or more of the techniques we have learned in class. 
+* After selecting a dataset and identifying the goal, write out a proposed analysis plan using template provided and submit it through GitHub for review.
+* Apply the techniques outlined and come up with a result for the dataset that you proposed.
+* Assemble a Jupyter notebook that communicates your hypothesis, methods, and results. Submit this as your final project.
+* Submit feedback about group and individual group members. This is done individually. 
+
+We strongly encourage you to discuss potential project ideas with your TAs and IAs, and/or with Prof. Ellis! This will give us a chance to make sure you’re on the right track even before you submit your draft. 
+
+
+### Taboo topics
+
+While you are *encouraged* to pick a topic that helps our understanding of the world/society, you are free to pick *almost any* topic for your final project. That said, there are a few off-limit (or "taboo") topics. These topics are off-limits for three reasons: (1) these projects have been done a whole bunch of times on Kaggle already, which (2) limits the amount you can learn, and (3) makes these projects super boring to read and grade. We encourage you to pick a topic you care about, not one that is the easiest to find on the Internet. 
+
+The taboo topics in COGS 108 are:
+
+- movie recommendation system
+- youtube video comments / trending analysis
+- Kickstarter success prediction/analysis
+- prediction of what makes a song popular on Spotify 
+
+If you think you have a novel idea and *really, really* want to work on any of these  topics, reach out to Prof Ellis, and she'll let you know if your idea will work.
+
+The following datasets from kaggle are also off limits as your main dataset*:
+
+  - [Trending Youtube Video Statistics](https://www.kaggle.com/datasnaek/youtube-new)
+  - [120 Years of Olympic history: athletes and results](https://www.kaggle.com/heesoo37/120-years-of-olympic-history-athletes-and-results)
[…]
