+++ instructions.md
@@ -25,13 +25,21 @@ tool, which is a requirement for running the Makefile.
 
 ## Content
 
-Content for the website is copied in from the COGS108 organization.
+Content for the website is copied in from the
+[COGS108 organization](https://github.com/cogs108).
 
-Note that none of sources for the materials are managed or stored in this repository.
+The idea here is that this repository copies in materials created for COGS108, and hosts a public version.
 
-Materials should not be edited directly in this repository, other than to edit out class details for the public version.
+Note that as this procedure copies in files from the course organization, the original source
+files for these materials are not stored in this repository. Any development of these materials
+should be done on the original files in the COGS108 organization.
 
-All development of materials must be done on the COGS108 organization.
+This repository does contain copies of the materials, as are presented on the website.
+Note that these files are likely to have been edited for public release, and so may differ
+from the original files.
+
+Note that for rebuilding the public website, there are manual edits to some files that are needed,
+as described below.
 
 #### Tutorials
 
@@ -45,13 +53,18 @@ Assignment materials are copied from the [Demo Assignments](https://github.com/C
 
 Assignments need to be updated to replace the header of the assignments to indicate that they are not valid assignments for 108.
 
-Currently, this is semi-automated, with scripts in development to update assignments programmatically, and some manually tuning needed.
+Currently, the Makefile will copy over and rename files the DX (for "demo"). Manual tuning is then needed to remove course information and add a note that these are not valid 108 files.
+
+Note that with any updates, if any data files change, this needs to be coordinated with the
+[Data](https://github.com/dataScienceInPractice/Data) repository.
 
 #### Project
 
-Project materials are copied over from the [Projects](https://github.com/COGS108/Projects) repository.
+Project materials are copied over from the
+[Projects](https://github.com/COGS108/Projects) repository, and from the
+[group template](https://github.com/COGS108/group_template) repository.
 
-The project file needs editing, to remove sections are specific to the course. Currently, this is done manually.
+The project files need manual editing, to remove sections are specific to the course.
 
 ## Build Schedule
 
