### 2022-02-21 9d7efce
fix project page titles
