### 2022-02-21 8b7dc7c
minor paper updates
