+++ paper/paper.bib
@@ -7,34 +7,28 @@
     number = {11},
     journal = {Journal of Open Source Education},
     author = {Jupyter, Project and Blank, Douglas and Bourgin, David and Brown, Alexander and Bussonnier, Matthias and Frederic, Jonathan and Granger, Brian and Griffiths, Thomas and Hamrick, Jessica and Kelley, Kyle and Pacer, M and Page, Logan and PÃ©rez, Fernando and Ragan-Kelley, Benjamin and Suchow, Jordan and Willing, Carol},
-    month = jan,
     year = {2019},
 }
 
-@software{chris_holdgraf_2019_2799972,
-  author       = {Chris Holdgraf and
-                  Jan Kleinert and
-                  Elizabeth DuPre and
-                  Mainak Jas and
-                  Alexander Morley and
-                  Matthew Brett and
-                  Matt Craig and
-                  Erik Sundell and
-                  Sam Lau and
-                  Luke and
-                  gaow and
-                  stafforddavidj and
-                  cnydw and
-                  Zachary Sailer and
-                  Tom and
-                  Mathieu Boudreau and
-                  James Mason and
-                  Ariel Rokem},
-  title        = {jupyter/jupyter-book: v0.5},
-  month        = may,
-  year         = 2019,
+@software{executable_books_community_2020_4539666,
+  author       = {Executable Books Community},
+  title        = {Jupyter Book},
+  year         = 2020,
   publisher    = {Zenodo},
-  version      = {v0.5},
-  doi          = {10.5281/zenodo.2799972},
-  url          = {https://doi.org/10.5281/zenodo.2799972}
+  version      = {v0.10},
+  doi          = {10.5281/zenodo.4539666},
+  url          = {https://doi.org/10.5281/zenodo.4539666}
+}
+
+@article{donoghue_teaching_2020,
+  title = {Teaching {Creative} and {Practical} {Data} {Science} at {Scale}},
+  volume = {29},
+  issn = {1069-1898},
+  url = {https://www.tandfonline.com/doi/full/10.1080/10691898.2020.1860725},
+  doi = {10.1080/10691898.2020.1860725},
+  number = {sup1},
+  journal = {Journal of Statistics Education},
+  author = {Donoghue, Thomas and Voytek, Bradley and Ellis, Shannon E.},
+  year = {2021},
+  pages = {S27--S39},
 }
