+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-Data Science in Practice is a collection of openly available materials, including tutorials and assignments, for learning how to integrate the many skills of data science. The course materials focus on the day-to-day practicalities of hands-on data science, with a particular emphasis on gaining a working familiarity with real-world applications and gaining a 'data intuition'. This collection of materials was originally developed for a course at UC San Diego, but have been updated and publicly released, and are available at: https://datascienceinpractice.github.io/
+Data Science in Practice is a collection of openly available materials, including tutorials and assignments, for learning how to integrate the many skills of data science. The course materials focus on the day-to-day practicalities of hands-on data science, with a particular emphasis on gaining a working familiarity with real-world applications and gaining a 'data intuition'. This collection of materials was originally developed for a course at UC San Diego designed to teach creative and practical data science at scale [@donoghue_teaching_2020]. The materials for this course have been updated and made publicly available, and are hosted at: https://datascienceinpractice.github.io/.
 
 Topics covered in the data science in practice tutorials include:
 
@@ -44,7 +44,7 @@ These topics are further explored in available assignments, which cover:
 * Collecting web data, applying data protection policies, anonymizing data, and adversarial attacks for deanonymizing data.
 * Data analyses, including statistical analyses, applying linear models, and creating visualizations.
 
-The materials are developed in the Python (>= 3.6) programming language, using a standard collection of packages in the scientific Python environment, which can be installed using the Anaconda distribution. All materials are built as Jupyter notebooks, with the assignments being built with the nbgrader extension [@jupyter_nbgrader_2019]. All the materials are hosted online, using the Jupyter Book tool [@chris_holdgraf_2019_2799972], from which all the source notebooks can be downloaded to be run locally.
+The materials are developed in the Python (>= 3.6) programming language, using a standard collection of packages in the scientific Python environment, which can be installed using the Anaconda distribution. All materials are built as Jupyter notebooks, with the assignments being built with the nbgrader extension [@jupyter_nbgrader_2019]. All the materials are hosted online, using the Jupyter Book tool [@executable_books_community_2020_4539666], from which all the source notebooks can be downloaded to be run locally.
 
 # Statement of Need
 
@@ -52,11 +52,11 @@ The field of data science has been rapidly expanding, creating a need for access
 
 There are many available resources for topics within and related to data science, including dedicated tutorials for data science tools and software packages. What can still be difficult, for the novice, is learning how to find and navigate through these materials. A key goal of this course and these materials is to offer a curated introduction to the many topics and available tools, and some initial guided work to make sure users can start to engage with the many aspects of data science. Throughout the course materials, there are many links to other resources. The goal is that these materials be a starting place for the potential user, and a launching off point to the many other more specific resources and tutorials available.
 
-Data science is an interdisciplinary field, requiring expertise from across a range of relevant fields - including technical aspects such as software, computation, statistics, mathematics and machine learning, as well as topics such as research design, contextual understanding of data, ethics, and an understanding of the potential impacts. These materials aim to encompass these multiple elements of data science, focusing not only on the technical aspects of doing data science, but also acknowledging and emphasizing the social impacts and responsibilities of practicing data scientists. These materials are part of an emerging field of integrated data science, as compared to some more traditional courses and materials that focus on, for example, more detailed machine learning or computation.
[…]
