### 2022-02-21 DataScienceInPractice/Site#4 @TomDonoghue
Thanks for the note @krother - I have now made a release tag for the current version of the project:
https://github.com/DataScienceInPractice/Site/releases/tag/1.0
