### 2022-02-21 #121 @TomDonoghue
Hey all! I have come back around to this and done a round of edits on the project. As well as responding to current reviewer comments, I fixed up a couple things on the site (including fixing up the project section), and also make some minor updates to the paper (re-generated version available above).

Thank you to @krother for reviewing this work, and for suggesting some helpful comments! I believe I have updated everything related to your review, but please let me know if I missed anything.

@KiCh84 - is there anything else you need from us to help you finish the review? I see that you've checked off most of the comments - please let us know if you have any comments or suggestions for helping to get the remaining topics checked off.

@magsol - thanks for keeping track of this, and sorry I was a little slow at getting back to it recently. I believe we have currently done everything on our side, and are just waiting on the final reviewer checks - please let us know if there is anything else we can do to finish this one up!
