### 2022-05-19 #121 @magsol
@TomDonoghue At this point, I think it'd be easiest for me to go through the outstanding points on @KiCh84's reviewer checklist. Fortunately, most of the items are finished; the unresolved ones seem to revolve around the paper itself. I should have some time tomorrow to go through them and then, providing everything checks out, we can start the publication process.

I sincerely apologize for how long it's taken. Let's see if we can't get this wrapped by the weekend.
