### 2022-05-19 #121 @TomDonoghue
@magsol - thank you, that sounds like a practical and efficient way to try and wrap this up! Please let us know of any comments and suggestions for the paper / project, and I will try to address them as soon as I can!
