### 2022-05-20 #121 @TomDonoghue
Thanks @magsol - for the quick checks here!

I have added Zenodo to the org / repo in order to create an archive DOI:
Here's the Zenodo link: https://zenodo.org/record/6568091
Here's the DOI: 10.5281/zenodo.6568091

Note: I had to create a new release to start Zenodo tracking (or, at least, I wasn't sure how to retro-actively add a Zenodo DOI to a previous release), so there is now a 1.1 release to kickstart this, though there are no content changes to what is tagged under the 1.0 release (see release page: https://github.com/DataScienceInPractice/Site/releases). I don't think this matters in any meaningful way, just wanted to note it in case it's important to update a version number to 1.1 somewhere, since the project was previously tagged to 1.0.
