### 2022-05-20 #121 @labarba
We request that authors edit the metadata of the Zenodo deposit so title and author list match the JOSE paper. It's just cleaner that way as readers see these as part of the "same scholarly object." Could you do that? (Note that you may not want Zenodo to do _automatic_ updates of versions with each release. Note also that Zenodo pulls every committer into the author list.)
