### 2022-05-20 #121 @TomDonoghue
I've updated the Zenodo record to match the title of the JOSE paper, and checked that the authors match.
This should now be updated: https://zenodo.org/record/6568091
