### 2022-05-20 #121 @labarba
- [ ] for the JupyterBooks reference, please add `{}` around the author `Executable Books Community` so that it is rendered like that, and not like "Community, E. B."
