### 2022-05-20 #121 @labarba
- [ ] on [line 59](https://github.com/DataScienceInPractice/Site/blob/8b7dc7ce4bda75bf234363b69f4b7c6be30acf6e/paper/paper.md?plain=1#L59): "This set of materials were" >> was
