### 2022-05-23 3d0be8c
wording & ref fixes
