+++ paper/paper.bib
@@ -6,12 +6,12 @@
     doi = {10.21105/jose.00032},
     number = {11},
     journal = {Journal of Open Source Education},
-    author = {Jupyter, Project and Blank, Douglas and Bourgin, David and Brown, Alexander and Bussonnier, Matthias and Frederic, Jonathan and Granger, Brian and Griffiths, Thomas and Hamrick, Jessica and Kelley, Kyle and Pacer, M and Page, Logan and PÃ©rez, Fernando and Ragan-Kelley, Benjamin and Suchow, Jordan and Willing, Carol},
+    author = {{Jupyter Project} and Blank, Douglas and Bourgin, David and Brown, Alexander and Bussonnier, Matthias and Frederic, Jonathan and Granger, Brian and Griffiths, Thomas and Hamrick, Jessica and Kelley, Kyle and Pacer, M and Page, Logan and PÃ©rez, Fernando and Ragan-Kelley, Benjamin and Suchow, Jordan and Willing, Carol},
     year = {2019},
 }
 
 @software{executable_books_community_2020_4539666,
-  author       = {Executable Books Community},
+  author       = {{Executable Books Community}},
   title        = {Jupyter Book},
   year         = 2020,
   publisher    = {Zenodo},
