+++ paper/paper.bib
@@ -1,13 +1,13 @@
 @article{jupyter_nbgrader_2019,
-    title = {nbgrader: {A} {Tool} for {Creating} and {Grading} {Assignments} in the {Jupyter} {Notebook}},
-    volume = {2},
-    issn = {2577-3569},
-    url = {https://jose.theoj.org/papers/10.21105/jose.00032},
-    doi = {10.21105/jose.00032},
-    number = {11},
-    journal = {Journal of Open Source Education},
-    author = {{Jupyter Project} and Blank, Douglas and Bourgin, David and Brown, Alexander and Bussonnier, Matthias and Frederic, Jonathan and Granger, Brian and Griffiths, Thomas and Hamrick, Jessica and Kelley, Kyle and Pacer, M and Page, Logan and PÃ©rez, Fernando and Ragan-Kelley, Benjamin and Suchow, Jordan and Willing, Carol},
-    year = {2019},
+  title = {nbgrader: {A} {Tool} for {Creating} and {Grading} {Assignments} in the {Jupyter} {Notebook}},
+  volume = {2},
+  issn = {2577-3569},
+  url = {https://jose.theoj.org/papers/10.21105/jose.00032},
+  doi = {10.21105/jose.00032},
+  number = {11},
+  journal = {Journal of Open Source Education},
+  author = {{Jupyter Project} and Blank, Douglas and Bourgin, David and Brown, Alexander and Bussonnier, Matthias and Frederic, Jonathan and Granger, Brian and Griffiths, Thomas and Hamrick, Jessica and Kelley, Kyle and Pacer, M and Page, Logan and PÃ©rez, Fernando and Ragan-Kelley, Benjamin and Suchow, Jordan and Willing, Carol},
+  year = {2019},
 }
 
 @software{executable_books_community_2020_4539666,
@@ -27,7 +27,7 @@
   url = {https://www.tandfonline.com/doi/full/10.1080/10691898.2020.1860725},
   doi = {10.1080/10691898.2020.1860725},
   number = {sup1},
-  journal = {Journal of Statistics Education},
+  journal = {Journal of Statistics and Data Science Education},
   author = {Donoghue, Thomas and Voytek, Bradley and Ellis, Shannon E.},
   year = {2021},
   pages = {S27--S39},
